feat(GeneratePalette.jsx): add response_format parameter to API request to receive response in JSON format
The response_format parameter is added to the API request in order to specify that the desired response format is a JSON object. This change is made to ensure that the response from the API is received in the expected format, allowing for easier handling and processing of the data.